requirements.txt
flask
pymongo[srv]
requests
flask-restful
flask-cors
beautifulsoup4
python-dotenv
gunicorn
apscheduler